refactor shop and item into different directories
