+++ app/item.js
@@ -1,0 +1,9 @@
+class Item {
+  constructor(name, sellIn, quality){
+    this.name = name;
+    this.sellIn = sellIn;
+    this.quality = quality;
+  }
+}
+
+module.exports = Item;
